+++ frontend/src/App.tsx
@@ -1,9 +1,7 @@
-import { Button, Header, Home, UsersList, Match, Messages } from "@/Components.tsx";
-import { useState } from "react";
-import { Link, Route, Routes, Router, BrowserRouter } from "react-router-dom";
-import reactLogo from "@images/react.svg";
-import viteLogo from "/vite.svg";
+import { Home } from "@/Components/HomePage.tsx";
+import { Link, Route, Routes, BrowserRouter } from "react-router-dom";
 import "@css/App.css";
+import { Match } from "@/Components/Match.tsx";
 
 // This is our first React "Component"
 export function App() {
