+++ frontend/src/Components/ExampleButton.tsx
@@ -1,0 +1,24 @@
+/*
+ ** Hooks—functions starting with `use`—can only be called at the top level of your components
+ * or [your own Hooks.](https://beta.reactjs.org/learn/reusing-logic-with-custom-hooks)**
+ * You can’t call Hooks inside conditions, loops, or other nested functions. Hooks are functions,
+ * but it’s helpful to think of them as unconditional declarations about your component’s needs.
+ * You “use” React features at the top of your component similar to how you “import” modules
+ * at the top of your file.
+ */
+import { useState } from "react";
+
+export const ExampleButton = () => {
+	const [clicks, setClicks] = useState(0);
+
+	return (
+		<button
+			onClick={() => {
+				console.log("Clicked!");
+				setClicks(clicks + 1);
+			}}
+		>
+			Clicks: {clicks}
+		</button>
+	);
+};
