+++ frontend/src/Components/Match.tsx
@@ -1,0 +1,5 @@
+export const Match = () => {
+	return (
+		<div>"MATCH PAGE"</div>
+	);
+};
